updated model calibration, added geothermal gradient plot to figure in main nb

+++ lib/heat_flow_model.py
@@ -111,8 +111,8 @@ def interpolate_data(xyz_array, data, dx, dy, limit_number_of_nodes=True, max_no
     if (len(xi) * len(yi)) > max_nodes:
         print('warning, interpolating data on raster with >1e6 nodes')
         if limit_number_of_nodes is True:
-            xi = np.linspace(xyz_array[:, 0].min(), xyz_array[:, 0].max(), np.sqrt(max_nodes))
-            yi = np.linspace(xyz_array[:, 1].min(), xyz_array[:, 1].max(), np.sqrt(max_nodes))
+            xi = np.linspace(xyz_array[:, 0].min(), xyz_array[:, 0].max(), int(np.sqrt(max_nodes)))
+            yi = np.linspace(xyz_array[:, 1].min(), xyz_array[:, 1].max(), int(np.sqrt(max_nodes)))
 
     xg, yg = np.meshgrid(xi, yi)
     xgf, ygf = xg.flatten(), yg.flatten()
